move Biospecimen table to explore package
