+++ components/AtlasTable.tsx
@@ -2,8 +2,7 @@ import _ from 'lodash';
 import { NextRouter } from 'next/router';
 import Tooltip from 'rc-tooltip';
 import React from 'react';
-import { getDefaultDataTableStyle } from '../lib/dataTableHelpers';
-import { Atlas, setTab } from '../lib/helpers';
+import { setTab } from '../lib/helpers';
 import { observer } from 'mobx-react';
 import { action, computed, makeObservable, observable } from 'mobx';
 import classNames from 'classnames';
@@ -16,7 +15,8 @@ import { PublicationPageLink, PUBLICATIONS } from '../lib/publications';
 
 import { ISelectedFiltersByAttrName } from '../packages/data-portal-filter/src/libs/types';
 import EnhancedDataTable from '../packages/data-portal-table/src/components/EnhancedDataTable';
-import { Entity } from '../packages/data-portal-commons/src/libs/entity';
+import { Atlas, Entity } from '../packages/data-portal-commons/src/libs/entity';
+import { getDefaultDataTableStyle } from '../packages/data-portal-table/src/libs/helpers';
 
 interface IAtlasTableProps {
     router: NextRouter;
