+++ packages/data-portal-table/src/libs/helpers.tsx
@@ -52,3 +52,18 @@ export function resolveColumnVisibility(
 
     return colVis;
 }
+export function getDefaultDataTableStyle() {
+    return {
+        headCells: {
+            style: {
+                fontSize: 16,
+                fontWeight: 'bold',
+            },
+        },
+        cells: {
+            style: {
+                fontSize: 14,
+            },
+        },
+    };
+}
